# optimize window function's `find_ranges_in_range`